Accion para enviar un email

+++ database/migrations/2020_07_28_204143_create_expenses_table.php
@@ -15,11 +15,11 @@ class CreateExpensesTable extends Migration
     {
         Schema::create('expenses', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('expense_report_id');
+            $table->unsignedBigInteger('expense_report_id');
             $table->text('description');
             $table->decimal('amount');
             $table->timestamps();
-
+            
             $table->foreign('expense_report_id')->references('id')->on('expense_reports');
         });
     }
